Move also AreenaApiProgramInfo, EpisodeMetadata
